% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pirate.R
\name{geom_pirate}
\alias{geom_pirate}
\title{Pirate plots}
\usage{
geom_pirate(
  mapping = NULL,
  data = NULL,
  ...,
  points = TRUE,
  bars = TRUE,
  lines = TRUE,
  cis = TRUE,
  violins = TRUE,
  points_params = list(shape = 1, size = 1),
  bars_params = list(alpha = 0.3, width = 0.9),
  lines_params = list(size = 0.5, width = 0.9),
  cis_params = list(fill = "white", size = 0.5, alpha = 0.5, width = 0.8),
  violins_params = list(fill = "white", size = 0.5, alpha = 0, width = 0.7),
  jitter_width = 0.2,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{points}{logical indicating whether to show points}

\item{bars}{logical indicating whether to show mean bars}

\item{lines}{logical indicating whether to show mean lines}

\item{cis}{logical indicating whether to show 95\% confidence intervals boxes}

\item{violins}{logical indicating whether to show violins}

\item{points_params}{list of parameters to pass to points layer}

\item{bars_params}{list of parameters to pass to bars layer}

\item{lines_params}{list of parameters to pass to lines layer}

\item{cis_params}{list of parameters to pass to CIs layer}

\item{violins_params}{list of parameters to pass to violins layer}

\item{jitter_width}{amount of horizontal jitter added to the locations of the
points (defaults to 20\% of the resolution of the data)}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical indicating whether this layer be included in the
legends (defaults to FALSE)}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
A pirate plot
(\url{https://cran.r-project.org/web/packages/yarrr/vignettes/pirateplot.html})
is a way of displaying data where a continuous dependent variable is a
function of a categorical independent variable, in a more informative way
than the traditional barplot. \code{geom_pirate} plots the raw data as points
(using \link[ggplot2]{geom_jitter}), along with layers showing descriptive
and inferential statistics -- bars indicating means (using
\link[ggplot2]{geom_col}), horizontal line indicating means (using
\link[ggplot2]{geom_crossbar}), boxes indicating 95\% confidence intervals
assuming a normal sampling distribution (using \link[ggplot2]{geom_tile}),
and violins indicating the density (using \link[ggplot2]{geom_violin}).
}
\examples{
ggplot(mpg, aes(x = class, y = cty)) +
  geom_pirate(aes(colour = class))
}
