% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pirate.R
\name{geom_pirate}
\alias{geom_pirate}
\title{Pirate plots}
\usage{
geom_pirate(mapping = NULL, data = NULL, ..., points = TRUE,
  bars = TRUE, lines = TRUE, cis = TRUE, violins = TRUE,
  width_points = 0.1, width_bars = 0.9, width_lines = 0.9,
  width_cis = 0.8, width_violins = 0.7, alpha_points = 1,
  alpha_bars = 0.3, alpha_cis = 0.5, alpha_violins = 0, trim = TRUE,
  scale = "area", fatten = 2.5, na.rm = FALSE, show.legend = FALSE,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{points}{logical indicating whether to show points}

\item{bars}{logical indicating whether to show mean bars}

\item{lines}{logical indicating whether to show mean lines}

\item{cis}{logical indicating whether to show confidence intervals boxes}

\item{violins}{logical indicating whether to show violins}

\item{width_points}{Amount of horizontal jitter added to the locations of the
points. Defaults to 10\% of the resolution of the data.}

\item{width_bars}{Width of mean bars. Defaults to 90\% of the resolution of
the data.}

\item{width_lines}{Width of mean lines. Defaults to 90\% of the resolution of
the data.}

\item{width_cis}{Width of confidence interval boxes. Defaults to 80\% of the
resolution of the data.}

\item{width_violins}{Width of violins. Defaults to 70\% of the resolution of
the data.}

\item{alpha_points}{Opacity of the points. Defaults to 1 (not transparent at
all).}

\item{alpha_bars}{Opacity of the mean bars. Defaults to 0.3.}

\item{alpha_cis}{Opacity of the confidence interval boxes. Defaults to 0.5.}

\item{alpha_violins}{Opacity of the violins. Defaults to 0 (completely
transparent).}

\item{trim}{If \code{TRUE} (default), trim the tails of the violins
to the range of the data. If \code{FALSE}, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{fatten}{A multiplicative factor used to increase the size of the mean
lines.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical indicating whether this layer be included in the
legends? NA includes if any aesthetics are mapped. FALSE (the default)
never includes, and TRUE always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\description{
A pirate plot
(\url{https://cran.r-project.org/web/packages/yarrr/vignettes/pirateplot.html})
is a way of displaying data where a continuous dependent variable is a
function of a categorical independent variable, in a more informative way
than the traditional barplot. \code{geom_pirate} plots the raw data as points
(using \link[ggplot2]{geom_jitter}), along with layers showing descriptive
and inferential statistics -- bars indicating means (using
\link[ggplot2]{geom_col}), horizontal line indicating means (using
\link[ggplot2]{geom_crossbar}), boxes indicating 95\% confidence intervals
assuming a normal sampling distribution (using \link[ggplot2]{geom_rect}),
and violins indicating the density (using \link[ggplot2]{geom_violin}).
}
\examples{
ggplot(mpg, aes(x = class, y = displ)) +
  geom_pirate()

ggplot(mpg, aes(x = class, y = displ)) +
  geom_pirate(aes(colour = class, fill = class))
}
